Alter initDb.ts to export init function so test-db may also be initialised with production tables
