+++ src/db/testUtils.ts
@@ -1,0 +1,7 @@
+import getDb from ".";
+
+export const createTable = async function (tableName: string) {
+  const db = getDb();
+
+  return db.query(`CREATE TABLE IF NOT EXISTS ${tableName} (id SERIAL PRIMARY KEY, name VARCHAR(40) not null, price NUMERIC(10, 2));`)
+}
